# Conflicts:
#	src/app/components/user/user.component.ts
#	src/assets/i18n/ca.json

+++ src/app/components/layers/layers.component.ts
@@ -4,6 +4,8 @@ import { UtilsService } from '../../services/utils.service';
 import { Router } from '@angular/router';
 import { environment } from 'src/environments/environment';
 import { Subject } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogMessageComponent } from 'dist/sitmun-frontend-gui/';
 
 @Component({
   selector: 'app-layers',
@@ -16,7 +18,8 @@ export class LayersComponent implements OnInit {
   themeGrid: any = environment.agGridTheme;
   columnDefs: any[];
 
-  constructor(public cartographyService: CartographyService,
+  constructor(public dialog: MatDialog,
+    public cartographyService: CartographyService,
     private utils: UtilsService,
     private router: Router,
   ) {
@@ -91,13 +94,23 @@ export class LayersComponent implements OnInit {
   }
 
   removeData(data: Cartography[]) {
-    const promises: Promise<any>[] = [];
-    data.forEach(cartography => {
 
-      promises.push(new Promise((resolve, reject) => {​​​​​​​ this.cartographyService.delete(cartography).toPromise().then((resp) =>{​​​​​​​resolve()}​​​​​​​)}​​​​​​​));
-      Promise.all(promises).then(() => {
-        this.dataUpdatedEvent.next(true);
-      });
+    const dialogRef = this.dialog.open(DialogMessageComponent);
+    dialogRef.componentInstance.title=this.utils.getTranslate("Caution");
+    dialogRef.componentInstance.message=this.utils.getTranslate("RemoveMessage");
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        if(result.event==='Accept') {  
+          const promises: Promise<any>[] = [];
+          data.forEach(cartography => {
+            promises.push(new Promise((resolve, reject) => {​​​​​​​ this.cartographyService.delete(cartography).toPromise().then((resp) =>{​​​​​​​resolve()}​​​​​​​)}​​​​​​​));
+            Promise.all(promises).then(() => {
+              this.dataUpdatedEvent.next(true);
+            });
+          });
+      
+       }
+      }
     });
 
   }
