+++ src/app/components/service/service.component.ts
@@ -4,6 +4,8 @@ import { UtilsService } from '../../services/utils.service';
 import { Router } from '@angular/router';
 import { environment } from 'src/environments/environment';
 import { Subject } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogMessageComponent } from 'dist/sitmun-frontend-gui/';
 
 @Component({
   selector: 'app-service',
@@ -16,7 +18,8 @@ export class ServiceComponent implements OnInit {
   themeGrid: any = environment.agGridTheme;
   columnDefs: any[];
 
-  constructor(public serviceService: ServiceService,
+  constructor(public dialog: MatDialog,
+    public serviceService: ServiceService,
     private utils: UtilsService,
     private router: Router,
   ) {
@@ -87,13 +90,26 @@ export class ServiceComponent implements OnInit {
   }
 
   removeData(data: Service[]) {
-    const promises: Promise<any>[] = [];
-    data.forEach(service => {
-      promises.push(new Promise((resolve, reject) => {​​​​​​​ this.serviceService.delete(service).toPromise().then((resp) =>{​​​​​​​resolve()}​​​​​​​)}​​​​​​​));
-      Promise.all(promises).then(() => {
-        this.dataUpdatedEvent.next(true);
-      });
+
+    
+    const dialogRef = this.dialog.open(DialogMessageComponent);
+    dialogRef.componentInstance.title=this.utils.getTranslate("Caution");
+    dialogRef.componentInstance.message=this.utils.getTranslate("RemoveMessage");
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        if(result.event==='Accept') {  
+          const promises: Promise<any>[] = [];
+          data.forEach(service => {
+            promises.push(new Promise((resolve, reject) => {​​​​​​​ this.serviceService.delete(service).toPromise().then((resp) =>{​​​​​​​resolve()}​​​​​​​)}​​​​​​​));
+            Promise.all(promises).then(() => {
+              this.dataUpdatedEvent.next(true);
+            });
+          });
+       }
+      }
     });
+
+
 
   }
